refactor bounty sub commands
