+++ src/app/service/bounty/claim.ts
@@ -1,0 +1,3 @@
+export default async (): Promise<any> => {
+	return;
+};
