+++ src/app/service/bounty/list.ts
@@ -1,0 +1,3 @@
+export default async (): Promise<any> => {
+	return;
+};
